ci: added a check to flag deprecated functions, fix: moved volunteer actions.ts file
