+++ eslint.config.mjs
@@ -17,10 +17,17 @@ const eslintConfig = [
 			"no-unused-vars": "off",
 			"@typescript-eslint/no-unused-vars": "off",
 			"@typescript-eslint/no-explicit-any": "off",
+			'@typescript-eslint/no-deprecated': 'error',
 		},
 	}),
 	{
-		ignores: [".next/**", "lint-staged.config.mjs"]
+		ignores: [".next/**", "lint-staged.config.mjs"],
+		languageOptions: {
+			parserOptions: {
+			  projectService: true,
+			  tsconfigRootDir: import.meta.dirname,
+			},
+		},
 	}
 ];
