+++ package.json
@@ -27,6 +27,7 @@
 		"dotenv": "^16.5.0",
 		"jsonwebtoken": "^9.0.2",
 		"lucide-react": "^0.513.0",
+		"mantine-form-zod-resolver": "^1.3.0",
 		"motion": "^12.19.2",
 		"next": "^15.4.1",
 		"num-words": "^1.2.3",
